feature: add enums
