+++ vCardLib.Tests/ModelTests/HobbyTests.cs
@@ -1,7 +1,7 @@
 ﻿// Created by Bolorunduro Winner-Timothy on  11/22/2016 at 5:18 AM
 
 using NUnit.Framework;
-using vCardLib.Helpers;
+using vCardLib.Enums;
 using vCardLib.Models;
 
 namespace vCardLib.Tests.ModelTests
