+++ vCardLib/Enums/ContactType.cs
@@ -1,0 +1,15 @@
+﻿namespace vCardLib.Enums
+{
+    /// <summary>
+    /// Contact types
+    /// </summary>
+    public enum ContactType
+    {
+        Individual,
+        Group,
+        Organization,
+        Location,
+        Application,
+        Device
+    }
+}
