+++ vCardLib/Enums/GenderType.cs
@@ -1,0 +1,13 @@
+﻿namespace vCardLib.Enums
+{
+    /// <summary>
+    /// Gender types
+    /// </summary>
+    public enum GenderType
+    {
+        Male,
+        Female,
+        Other,
+        None
+    }
+}
