+++ vCardLib/Enums/Level.cs
@@ -1,0 +1,12 @@
+﻿namespace vCardLib.Enums
+{
+	/// <summary>
+	/// Various activity levels
+	/// </summary>
+	public enum Level
+	{
+		High,
+		Medium,
+		Low
+	}
+}
