+++ vCardLib/Enums/OverWriteOptions.cs
@@ -1,0 +1,11 @@
+﻿namespace vCardLib.Enums
+{
+    /// <summary>
+    /// Available write options
+    /// </summary>
+    public enum OverWriteOptions
+    {
+        Proceed,
+        Throw
+    }
+}
