+++ vCardLib/Enums/vCardVersion.cs
@@ -1,0 +1,13 @@
+﻿namespace vCardLib.Enums
+{
+    /// <summary>
+    /// vCard version
+    /// </summary>
+    // ReSharper disable once InconsistentNaming
+    public enum vCardVersion
+    {
+        V2,
+        V3,
+        V4
+    }
+}
