+++ vCardLib/Models/Interest.cs
@@ -1,4 +1,4 @@
-﻿using vCardLib.Helpers;
+﻿using vCardLib.Enums;
 
 namespace vCardLib.Models
 {
